org.lockss.plugin.base.BaseArchivalUnit *changed pause(int) to pause(long)
org.lockss.plugin.highwire.HighWireArchivalUnit
*removed casting for call to pause

+++ src/org/lockss/plugin/base/BaseArchivalUnit.java
@@ -197,7 +197,7 @@ public abstract class BaseArchivalUnit implements ArchivalUnit {
     return false;
   }
 
-  protected void pause(int milliseconds) {
+  protected void pause(long milliseconds) {
     try {
       Thread thread = Thread.currentThread();
       thread.sleep(milliseconds);
